requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
allennlp==2.10.1
altair==4.2.2
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.5
blis==0.7.9
boto3==1.26.75
botocore==1.29.75
cached-path==1.1.6
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
chart-studio==1.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.7
cvxpy==1.3.0
cycler==0.11.0
cymem==2.0.7
datasets==2.9.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
docker-pycreds==0.4.0
docutils==0.19
easyocr==1.6.2
ecos==2.0.12
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
fairscale==0.4.6
fastjsonschema==2.16.2
filelock==3.7.1
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
h5py==3.8.0
huggingface-hub==0.10.1
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
iniconfig==2.0.0
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kivymd==1.1.1
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.1
llvmlite==0.39.1
lmdb==1.4.0
Markdown==3.1.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.5
mizani==0.8.1
more-itertools==9.0.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netkit==3.1.10
networkx==3.0
ninja==1.11.1
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.5.4.60
osqp==0.6.2.post8
packaging==23.0
palettable==3.3.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pathy==0.10.1
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
plotly==5.13.0
plotnine==0.10.1
pluggy==1.0.0
preshed==3.0.8
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycparser==2.21
pydantic==1.8.2
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pynndescent==0.5.8
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.19.3
pytesseract==0.3.10
pytest==7.2.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-json-logger==2.0.4
pytorch-lightning==0.7.5
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qdldl==0.1.5.post3
qtconsole==5.4.0
QtPy==2.3.0
redis==4.5.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rsa==4.9
s3transfer==0.6.0
sacremoses==0.0.53
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.10.0
scs==3.2.2
seaborn==0.12.2
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.97
sentry-sdk==1.15.0
seqeval==1.2.2
setproctitle==1.3.2
shapely==2.0.1
shortuuid==1.0.11
simpletransformers==0.63.9
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.3.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.5
stack-data==0.6.2
statsmodels==0.13.5
streamlit==1.18.1
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2
tenacity==8.2.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==1.1.0
terminado==0.17.1
tesseract==0.1.3
thinc==8.0.17
threadpoolctl==3.1.0
tifffile==2023.2.28
tinycss2==1.2.1
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1
torchaudio==0.12.1+cu116
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.20.1
typer==0.4.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
umap-learn==0.5.3
ump==0.2.3
uri-template==1.2.0
urllib3==1.26.14
validators==0.20.0
virtualenv==20.19.0
wandb==0.12.21
wasabi==0.10.1
watchdog==2.2.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
wget==3.2
widgetsnbextension==4.0.5
wordcloud==1.8.2.2
xxhash==3.2.0
yarl==1.8.2
zipp==3.14.0